David_Notification.jpg
- Changed format a bit.
It seems like it would mostly work but there are a couple of code errors unless I'm missing something.
isCancel isn't declared... this code is in a method that already exists so it needs/needed to be commented out.
I will try to test it with a UI soon. More focused on Demo 1 immediate goals.
Is there a reason to do this method over push notifications? Does doing it this way require the app to be open and "on" in order to send notifications?